--- BE/Main_Server/package.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.21.0",
-    "mongodb": "^6.9.0",
-    "mongoose": "^8.7.0"
-  }
-}
