--- BE/RSS_Server/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-index.mjs
-test_rss.mjs
