--- BE/RSS_Server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "rss-downloader",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mongodb": "^6.9.0",
-    "node-cron": "^3.0.3",
-    "node-fetch": "^3.3.2",
-    "rss-downloader": "file:",
-    "xml2js": "^0.6.2"
-  }
-}
