+++ Frontend/public/manifest.json
