+++ Frontend/src/components/header/Header.js
@@ -103,7 +103,7 @@ const Header = () => {
 
   const fetchSearchResults = async () => {
     try {
-      const response = await fetch(`http://localhost:3001/search?q=${keyword}`);
+      const response = await fetch(`http://localhost:5000/api/search?q=${keyword}`);
       if (!response.ok) {
         throw new Error("Có lỗi xảy ra khi tìm kiếm");
       }
